% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{combine_results}
\alias{combine_results}
\title{Combine Results Function}
\usage{
combine_results(...)
}
\arguments{
\item{...}{Variable number of Metropolis-Hastings result lists.}
}
\value{
A list containing combined p-values, bar_beta estimates, and average acceptance rate.
}
\description{
This function combines the results from multiple Metropolis-Hastings simulations.
}
\examples{
result1 <- list()
result2 <- list()
result3 <- list()
combine_results(result1, result2, result3)
}
