% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.lcj}
\alias{dat.lcj}
\title{A Meta-Analysis on the Effect of Progressive Resistance Strength Training Exercise vs. Control}
\format{
A data frame containing 33 studies with the observed effect sizes and their within-study variances.
\describe{
\item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
\item{\code{s2}}{the within-study variance for each study.}
}
}
\source{
Liu CJ, Latham NK (2009). "Progressive resistance strength training for
improving physical function in older adults." \emph{Cochrane Database of Systematic Reviews},
\bold{3}. Art. No.: CD002759. \if{html}{\out{<\doi{10.1002/14651858.CD002759.pub2}>}}
}
\usage{
data("dat.lcj")
}
\description{
This meta-analysis serves as an example to illustrate function usage.
}
\keyword{dataset}
