% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.slf}
\alias{dat.slf}
\title{A Meta-Analysis on the Effect of Nicotine Gum for Smoking Cessation}
\format{
A data frame containing 56 studies with the observed effect sizes and
their within-study variances.
\describe{
\item{\code{y}}{the observed effect size for each collected study in the meta-analysis.}
\item{\code{s2}}{the within-study variance for each study.}
}
}
\source{
Stead LF, Perera R, Bullen C, Mant D, Hartmann-Boyce J, Cahill K, Lancaster T (2012).
"Nicotine replacement therapy for smoking cessation." \emph{Cochrane Database of Systematic Reviews},
\bold{11}. Art. No.: CD000146. \if{html}{\out{<\doi{10.1002/14651858.CD000146.pub4}>}}
}
\usage{
data("dat.slf")
}
\description{
This meta-analysis serves as an example to illustrate function usage.
}
\keyword{dataset}
