% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{frequency_pvalue}
\alias{frequency_pvalue}
\title{Calculate the p-value for Cochran's Q test}
\usage{
frequency_pvalue(hat_beta, hat_sigma_sq)
}
\arguments{
\item{hat_beta}{The original or the simulated estimated effects.}

\item{hat_sigma_sq}{The squared standard errors of the estimated effects.}
}
\value{
A numeric value representing the p-value for Cochran's Q test.
}
\description{
Calculate the p-value for Cochran's Q test
}
\examples{
hat_beta <- c(0.5, 1.0, 1.5)
hat_sigma_sq <- c(0.2, 0.3, 0.4)
frequency_pvalue(hat_beta, hat_sigma_sq)
}
