% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{g}
\alias{g}
\title{Function g by log transformation}
\usage{
g(bar_beta, hat_beta, hat_sigma_sq, k)
}
\arguments{
\item{bar_beta}{The estimated true underlying effect.}

\item{hat_beta}{The original or the simulated estimated effects.}

\item{hat_sigma_sq}{The squared standard errors of the estimated effects.}

\item{k}{The level of heterogeneity.}
}
\value{
Numeric value of the log-transformed function g.
}
\description{
This function calculates the log-transformed value of g.
}
\keyword{internal}
