% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{hotelling_t_squared}
\alias{hotelling_t_squared}
\title{Hotelling T-squared Statistic Function}
\usage{
hotelling_t_squared(beta_prime_H, hat_beta, hat_sigma_sq, df)
}
\arguments{
\item{beta_prime_H}{The mean of estimated true underlying effect.}

\item{hat_beta}{The original or the simulated estimated effects.}

\item{hat_sigma_sq}{The squared standard errors of the estimated effects.}

\item{df}{Degrees of freedom for the test.}
}
\value{
p-value for the Hotelling T-squared test.
}
\description{
This function calculates the Hotelling T-squared statistic for multivariate hypothesis testing.
}
\keyword{internal}
