% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{inverse_P_mis}
\alias{inverse_P_mis}
\title{Inverse Probability of Misclassification Function}
\usage{
inverse_P_mis(
  P_mis_val,
  lower = 1e-06,
  upper = 100,
  tol = 1e-10,
  max_iter = 10^15
)
}
\arguments{
\item{P_mis_val}{The value of misclassification probability for which the inverse is sought.}

\item{lower}{The lower bound for the bisection method.}

\item{upper}{The upper bound for the bisection method.}

\item{tol}{The tolerance level for the bisection method.}

\item{max_iter}{The maximum number of iterations for the bisection method.}
}
\value{
Numeric value of the approximate inverse function value of \eqn{P_{\text{mis}}}.
}
\description{
This function approximates the inverse function value of \eqn{P_{\text{mis}}(k)} using the bisection method.
}
\examples{
inverse_P_mis(0.05)
}
