% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{mean_Q}
\alias{mean_Q}
\title{Calculate the weighted mean of beta estimates}
\usage{
mean_Q(hat_beta, hat_sigma_sq)
}
\arguments{
\item{hat_beta}{The original or the simulated estimated effects.}

\item{hat_sigma_sq}{The squared standard errors of the estimated effects.}
}
\value{
A numeric value representing the weighted mean of the effect estimates.
}
\description{
Calculate the weighted mean of beta estimates
}
\examples{
hat_beta <- c(0.5, 1.0, 1.5)
hat_sigma_sq <- c(0.2, 0.3, 0.4)
mean_Q(hat_beta, hat_sigma_sq)

}
