% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{P_mis}
\alias{P_mis}
\title{Probability of Misclassification Function}
\usage{
P_mis(k)
}
\arguments{
\item{k}{A numeric value representing the level of heterogeneity.}
}
\value{
Numeric value of the probability of misclassification.
}
\description{
This function calculates the probability of misclassification for a given value of k.
}
\examples{
P_mis(1)
}
