% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{q}
\alias{q}
\title{Proposed Function q by log transformation}
\usage{
q(bar_beta_star, hat_beta, hat_sigma_sq, k)
}
\arguments{
\item{bar_beta_star}{The simulated value of bar_beta.}

\item{hat_beta}{The original or the simulated estimated effects.}

\item{hat_sigma_sq}{The squared standard errors of the estimated effects.}

\item{k}{The level of heterogeneity.}
}
\value{
Logarithm of the density of the normal distribution for bar_beta_star.
}
\description{
This function calculates logarithm of the proposed density used in Metropolis-Hastings (MH).
}
\keyword{internal}
