% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle Function}
\usage{
shuffle(xv, rep)
}
\arguments{
\item{xv}{Input vector to be shuffled.}

\item{rep}{Number of times to perform the shuffle.}
}
\value{
Shuffled vector.
}
\description{
This function randomly shuffles elements in a vector.
}
\examples{
shuffle(c(1, 2, 3, 4, 5), 2)
}
