% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{sim_batch}
\alias{sim_batch}
\title{Simulate Batch Contamination Function}
\usage{
sim_batch(bb_sd, k, bbar, m, bc_grp_num, sample_size, Gv, Batch, rnse)
}
\arguments{
\item{bb_sd}{Standard deviation for batch effects. This determines the variability of the batch effects in the contaminated groups.}

\item{k}{Level of heterogeneity, affecting the variability of the true underlying effect across groups.}

\item{bbar}{True underlying effect. This is the baseline effect size which is perturbed by heterogeneity.}

\item{m}{Number of groups to simulate.}

\item{bc_grp_num}{Number of groups that are batch-contaminated.}

\item{sample_size}{Number of samples per group.}

\item{Gv}{Matrix of group indicators; each row corresponds to a group and each column to a sample.}

\item{Batch}{Matrix of batch effects; structured similarly to the \code{Gv} matrix, with each row corresponding to a group.}

\item{rnse}{Standard deviation of the random noise added to the simulation.}
}
\value{
A vector containing the simulated data, specifically the estimated coefficients and their standard errors from linear models applied to each group.
}
\description{
This function simulates batch contamination in a dataset by incorporating batch effects and heterogeneity in group data.
}
\examples{
rnse <- 1
sample_size <- 5
m <- 1
bc_grp_num <- 1 # number of batch-contaminated groups
Gv <- t(sapply(1:m, function(x) rbinom(sample_size, 1, 0.4)))
Batch <- t(apply(Gv, 1, function(x) shuffle(x, rep=floor(sample_size*0.2))))
sim_batch(0.1, 1, 0.5, 1, bc_grp_num, sample_size, Gv, Batch, rnse)
}
