% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{T_Q}
\alias{T_Q}
\title{Modified Cochran’s Q statistic T_Q}
\usage{
T_Q(hat_beta, bar_beta, k, hat_sigma_sq)
}
\arguments{
\item{hat_beta}{The original or the simulated estimated effects.}

\item{bar_beta}{The estimated true underlying effect.}

\item{k}{The level of heterogeneity.}

\item{hat_sigma_sq}{The squared standard errors of the estimated effects.}
}
\value{
Numeric value of the test statistic T_Q.
}
\description{
This function calculates the modified Cochran’s Q statistic.
}
\examples{
T_Q(c(0.4, 0.6), 0.5, 1, c(0.1, 0.2))
}
