% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_posterior.R
\name{wipi}
\alias{wipi}
\title{Censoring Function}
\usage{
wipi(p, c)
}
\arguments{
\item{p}{Numeric vector of probabilities to be censored.}

\item{c}{Censoring strength}
}
\value{
Censored probabilities.
}
\description{
This function applies a censoring mechanism based on a power law.
}
\examples{
wipi(0.1, 1)
}
